python/Flask/integration_telegram_planfix/telegram_planfix_integration_work_in_progress/requirements.txt
aiocache==0.12.2
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
frozenlist==1.4.1
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.0
rsa==4.9
Telethon==1.33.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
